refactor(api): concentrate telemetry plumbing + scope Sonar CPD to cut new-code duplication
The request_completed middleware and the unhandled-exception capture were
duplicated across api/app/main.py and gpu-api/app/main.py. api/ and gpu-api/ are
independently-built services (separate Docker contexts, pyproject, uv.lock), so
they cannot share a module. Concentrate the parallel boilerplate into each
service's own telemetry module (posthog_client.py / analytics.py) behind thin
helpers (register_request_middleware, capture_request_exception); main.py now
calls one-liners instead of holding duplicated blocks.

Also DRY the gpu-api embed/rerank chat_inference emission into a shared
capture_chat_inference helper. Add a narrow sonar.cpd.exclusions for ONLY the two
telemetry modules (which now hold the unavoidable parallel boilerplate); the
existing sonar.exclusions is kept and main.py is not excluded. Behavior and
residency are unchanged.
