+++ .sonarcloud.properties
@@ -4,3 +4,5 @@ sonar.projectKey=finki-hub-chat-bot
 sonar.sources=.
 sonar.python.version=3.14
 sonar.exclusions=web/components/ui/**,web/components/ai-elements/**
+# Parallel telemetry boilerplate for two independently-deployed services (separate Docker build contexts preclude sharing).
+sonar.cpd.exclusions=api/app/utils/posthog_client.py,gpu-api/app/utils/analytics.py
