+++ gpu-api/app/main.py
@@ -1,5 +1,4 @@
 import logging
-import time
 from asyncio import gather, to_thread
 from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
@@ -9,9 +8,7 @@ from fastapi import FastAPI, Request, status
 from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import RequestValidationError
 from fastapi.responses import JSONResponse
-from starlette.datastructures import Headers
 from starlette.middleware.cors import CORSMiddleware
-from starlette.types import ASGIApp, Message, Receive, Scope, Send
 
 from app.api.embeddings import router as embeddings_router
 from app.api.health import router as health_router
@@ -21,9 +18,9 @@ from app.llms.bge_m3 import init_bge_m3_embedder
 from app.llms.reranker import init_reranker
 from app.utils.analytics import (
     capture,
-    capture_exception,
+    capture_request_exception,
     init_analytics,
-    safe_response_id,
+    register_request_middleware,
     shutdown_analytics,
 )
 from app.utils.exceptions import ModelNotReadyError
@@ -64,70 +61,6 @@ async def lifespan(app: FastAPI) -> AsyncGenerator[None]:
     shutdown_analytics()
 
 
-# Extreme-noise paths kept out of request_completed to control event volume.
-_SKIP_PATHS: frozenset[str] = frozenset(
-    {"/docs", "/redoc", "/openapi.json", "/favicon.ico"},
-)
-_SKIP_PREFIXES: tuple[str, ...] = ("/health",)
-
-
-def _request_outcome(status_code: int) -> str:
-    if status_code >= 500:
-        return "server_error"
-    if status_code >= 400:
-        return "client_error"
-    return "ok"
-
-
-class RequestTrackingMiddleware:
-    """Emit one ``request_completed`` PostHog event per HTTP request (metadata only).
-
-    A pure ASGI middleware, not BaseHTTPMiddleware, so it never buffers a streaming body:
-    it only reads the status off ``http.response.start`` and times the whole request. The
-    matched route TEMPLATE is reported (never the raw path), so ids never leak and route /
-    person cardinality stays bounded.
-    """
-
-    def __init__(self, app: ASGIApp) -> None:
-        self.app = app
-
-    async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
-        if scope["type"] != "http":
-            await self.app(scope, receive, send)
-            return
-        path = scope.get("path", "")
-        if path in _SKIP_PATHS or path.startswith(_SKIP_PREFIXES):
-            await self.app(scope, receive, send)
-            return
-
-        start = time.perf_counter()
-        status_code = 500
-
-        async def send_wrapper(message: Message) -> None:
-            nonlocal status_code
-            if message["type"] == "http.response.start":
-                status_code = message["status"]
-            await send(message)
-
-        try:
-            await self.app(scope, receive, send_wrapper)
-        finally:
-            route = scope.get("route")
-            template = getattr(route, "path", None) or path
-            capture(
-                safe_response_id(Headers(scope=scope).get("x-response-id"))
-                or "gpu-api",
-                "request_completed",
-                {
-                    "route": template,
-                    "method": scope.get("method", ""),
-                    "status_code": status_code,
-                    "duration_ms": round((time.perf_counter() - start) * 1000, 1),
-                    "outcome": _request_outcome(status_code),
-                },
-            )
-
-
 def make_app(settings: Settings) -> FastAPI:
     """
     Create and configure the FastAPI application.
@@ -155,10 +88,8 @@ def make_app(settings: Settings) -> FastAPI:
         expose_headers=settings.EXPOSE_HEADERS,
     )
 
-    # Added last so it sits outermost and times the whole request. A handled error (e.g.
-    # 422/503) arrives as a normal response we read the status off; an unhandled 500
-    # propagates here as an exception, so status_code stays 500 and we still emit in finally.
-    app.add_middleware(RequestTrackingMiddleware)
+    # Added last so it sits outermost and times the whole request.
+    register_request_middleware(app)
 
     app.include_router(embeddings_router)
     app.include_router(streams_router)
@@ -202,10 +133,7 @@ def make_app(settings: Settings) -> FastAPI:
         exc: Exception,
     ) -> JSONResponse:
         logger.exception("Unhandled exception")
-        capture_exception(
-            exc,
-            properties={"path": request.url.path, "method": request.method},
-        )
+        capture_request_exception(request, exc)
         return JSONResponse(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             content={"detail": "An unexpected internal server error occurred."},
